+++ packages/pensopay_android/android/src/main/kotlin/com/flutter/pensopay/networking/pensopayapi/pensopaylink/payments/PPRefundPaymentRequest.kt
@@ -8,12 +8,12 @@ import org.json.JSONObject
 class PPRefundPaymentRequest(params: PPRefundPaymentParameters): PPrequest<PPPayment>(Request.Method.POST, "/payment/${params.payment_id}/refund", params, PPPayment::class.java) {}
 
 
-class PPRefundPaymentParameters(payment_id: Int, amount: Double?): JSONObject() {
+class PPRefundPaymentParameters(payment_id: Int, amount: Int?): JSONObject() {
 
     // Required Properties
     var payment_id: Int = payment_id
 
     // Optional Properties
-    var amount: Double? = amount
+    var amount: Int? = amount
 
 }
